Refactor workgroup table
