+++ src/client/src/pages/settings/settingsPage.tsx
@@ -10,9 +10,9 @@ import { useAuth } from "../../auth/useBdmsAuth.tsx";
 import { TabPanel } from "../../components/tabs/tabPanel.tsx";
 import AboutSettings from "./aboutSettings";
 import AdminSettings from "./admin/adminSettings";
+import { UserAdministration } from "./admin/userAdministration.tsx";
 import { UserDetail } from "./admin/userDetail.tsx";
-import { UserTable } from "./admin/userTable.tsx";
-import { WorkgroupTable } from "./admin/workgroupTable.tsx";
+import { WorkgroupAdministration } from "./admin/workgroupAdministration.tsx";
 import EditorSettings from "./editorSettings.tsx";
 import { SettingsHeader } from "./settingsHeader.tsx";
 import TermSettings from "./termSettings";
@@ -39,12 +39,12 @@ export const SettingsPage = () => {
       tabsArray.unshift({
         label: t("workgroups"),
         hash: "workgroups",
-        component: <WorkgroupTable users={users} />,
+        component: <WorkgroupAdministration users={users} />,
       });
       tabsArray.unshift({
         label: t("users"),
         hash: "users",
-        component: <UserTable setSelectedUser={setSelectedUser} users={users} setUsers={setUsers} />,
+        component: <UserAdministration setSelectedUser={setSelectedUser} users={users} setUsers={setUsers} />,
       });
     }
